include BaseClasses_Import.hpp
+ a bit of cleanup and a new base class (mask on/off)
