+++ addons/compat_sog/compat_sog_wardrobe/CfgWardrobe.hpp
@@ -1,17 +1,4 @@
-
-
-class EGVAR(wardrobe,base);
-
-class EGVAR(wardrobe,base_H_visor_up);
-class EGVAR(wardrobe,base_H_visor_down);
-
-class EGVAR(wardrobe,base_U_sleeves_up);
-class EGVAR(wardrobe,base_U_sleeves_down);
-
-class EGVAR(wardrobe,base_H_goggles_on);
-class EGVAR(wardrobe,base_H_goggles_off);
-
-
+#include "\z\ace\addons\wardrobe\BaseClasses_Import.hpp"
 
 class CfgWeapons {
     // Base Classes
