+++ addons/wardrobe/BaseClasses_Import.hpp
@@ -1,0 +1,19 @@
+class EGVAR(wardrobe,base);
+
+class EGVAR(wardrobe,base_H_visor_up);
+class EGVAR(wardrobe,base_H_visor_down);
+
+class EGVAR(wardrobe,base_H_goggles_on);
+class EGVAR(wardrobe,base_H_goggles_off);
+
+class EGVAR(wardrobe,base_H_mask_on);
+class EGVAR(wardrobe,base_H_mask_off);
+
+class EGVAR(wardrobe,base_U_sleeves_up);
+class EGVAR(wardrobe,base_U_sleeves_down);
+
+class EGVAR(wardrobe,base_U_gloves_on);
+class EGVAR(wardrobe,base_U_gloves_off);
+
+class EGVAR(wardrobe,base_U_jacket_open);
+class EGVAR(wardrobe,base_U_jacket_closed);
