+++ addons/wardrobe/CfgWardrobe.hpp
@@ -1,12 +1,3 @@
-
-/*
-class CfgGlasses {
-    #include "Facewear.hpp"
-};
-*/
-
-// The new way :sigh:
-
 class ADDON {
     #include "BaseClasses.hpp"
     #include "Uniforms.hpp"
